class 7 Guardrails Assignment/pyproject.toml
[project]
name = "class-7-guardrails-homework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.5",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
]
